service/app/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
fastapi==0.79.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.2
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn[standard]
pydantic[email]
aiofiles
werkzeug
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]

